+++ src/application/index.tsx
@@ -1,12 +1,12 @@
 import { useEffect } from "react";
 import { createRoot } from "react-dom/client";
+import type { TypedDocumentNode } from "@apollo/client";
 import {
   ApolloClient,
   ApolloProvider,
   InMemoryCache,
   makeVar,
   gql,
-  TypedDocumentNode,
 } from "@apollo/client";
 import { getOperationName } from "@apollo/client/utilities";
 import { print } from "graphql/language/printer";
@@ -15,7 +15,7 @@ import { colorTheme, listenForThemeChange } from "./theme";
 import { App } from "./App";
 import { fragmentRegistry } from "./fragmentRegistry";
 
-import {
+import type {
   GetAllMutations,
   GetAllMutationsVariables,
   GetQueries,
@@ -23,7 +23,8 @@ import {
   WatchedMutation,
   WatchedQuery,
 } from "./types/gql";
-import { QueryInfo } from "../extension/tab/helpers";
+import type { QueryInfo } from "../extension/tab/helpers";
+import type { JSONObject } from "./types/json";
 
 const cache = new InMemoryCache({
   fragments: fragmentRegistry,
@@ -152,7 +153,7 @@ export const writeData = ({
 }: {
   queries: QueryInfo[];
   mutations: QueryInfo[];
-  cache: Record<string, unknown>;
+  cache: JSONObject;
 }) => {
   const filteredQueries = queries.map(getQueryData).filter(Boolean);